+++ src/AppServices/Complaints/ComplaintFilters.cs
@@ -1,4 +1,4 @@
-using Cts.AppServices.Complaints.Dto;
+using Cts.AppServices.Complaints.Dto.Query;
 using Cts.Domain.Entities.Complaints;
 using GaEpd.AppLibrary.Domain.Predicates;
 using System.Linq.Expressions;
