+++ src/AppServices/Complaints/Dto/Query/ComplaintPublicSearchDto.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Cts.AppServices.Complaints.Dto;
+namespace Cts.AppServices.Complaints.Dto.Query;
 
 public record ComplaintPublicSearchDto
 {
