+++ src/AppServices/Complaints/Dto/Query/ComplaintSearchEnums.cs
@@ -2,7 +2,7 @@ using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 using System.Text.Json.Serialization;
 
-namespace Cts.AppServices.Complaints.Dto;
+namespace Cts.AppServices.Complaints.Dto.Query;
 
 [JsonConverter(typeof(JsonStringEnumConverter))]
 public enum SortBy
