+++ src/WebApp/Pages/Staff/Complaints/EditAction.cshtml.cs
@@ -2,7 +2,7 @@
 using Cts.AppServices.ComplaintActions;
 using Cts.AppServices.ComplaintActions.Dto;
 using Cts.AppServices.Complaints;
-using Cts.AppServices.Complaints.Dto;
+using Cts.AppServices.Complaints.Dto.Query;
 using Cts.AppServices.Complaints.Permissions;
 using Cts.WebApp.Models;
 using Cts.WebApp.Platform.PageModelHelpers;
