+++ src/WebApp/Pages/Staff/Complaints/Index.cshtml
@@ -1,5 +1,5 @@
 @page "{handler?}"
-@using Cts.AppServices.Complaints.Dto
+@using Cts.AppServices.Complaints.Dto.Query
 @using Cts.WebApp.Pages.Shared.DisplayTemplates
 @using Cts.WebApp.Platform.Constants
 @model IndexModel
