+++ tests/AppServicesTests/Complaints/PublicSearch.cs
@@ -1,6 +1,6 @@
-﻿using Cts.AppServices.Attachments;
+using Cts.AppServices.Attachments;
 using Cts.AppServices.Complaints;
-using Cts.AppServices.Complaints.Dto;
+using Cts.AppServices.Complaints.Dto.Query;
 using Cts.AppServices.UserServices;
 using Cts.Domain.Entities.Complaints;
 using Cts.Domain.Entities.Concerns;
