+++ tests/WebAppTests/Pages/Staff/Complaints/DetailsPageGetTests.cs
@@ -1,5 +1,5 @@
 using Cts.AppServices.Complaints;
-using Cts.AppServices.Complaints.Dto;
+using Cts.AppServices.Complaints.Dto.Query;
 using Cts.AppServices.Complaints.Permissions;
 
 namespace WebAppTests.Pages.Staff.Complaints;
